+++ packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/next/extensions/resolver/AssetExtensions.kt
@@ -17,8 +17,8 @@ import kotlinx.coroutines.withContext
 suspend fun ContentResolver.queryAssetDisplayName(contentUri: Uri): String? =
   queryOne(contentUri, MediaStore.MediaColumns.DISPLAY_NAME, Cursor::getString)
 
-suspend fun ContentResolver.queryGetCreationTime(contentUri: Uri): Long? =
-  queryOne(contentUri, MediaStore.MediaColumns.DATE_TAKEN, Cursor::getLong)
+suspend fun ContentResolver.queryAssetCreationTime(contentUri: Uri): Long? =
+  queryOne(contentUri, MediaStore.Images.Media.DATE_TAKEN, Cursor::getLong)
 
 suspend fun ContentResolver.queryAssetModificationTime(contentUri: Uri): Long? =
   queryOne(contentUri, MediaStore.MediaColumns.DATE_MODIFIED, Cursor::getLong)